move all installable files to installable folder
