+++ src/extensionConsole/pages/blueprints/Status.tsx
@@ -19,8 +19,8 @@ import styles from "./Status.module.scss";
 
 import React from "react";
 import { Button } from "react-bootstrap";
-import { type InstallableViewItem } from "./blueprintsTypes";
-import useInstallableViewItemActions from "./useInstallableViewItemActions";
+import { type InstallableViewItem } from "../../../installables/blueprintsTypes";
+import useInstallableViewItemActions from "../../../installables/useInstallableViewItemActions";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import {
   faCheck,
