+++ src/extensionConsole/pages/blueprints/listView/ListItem.tsx
@@ -19,9 +19,9 @@ import styles from "./ListItem.module.scss";
 
 import React from "react";
 import SharingLabel from "@/extensionConsole/pages/blueprints/labels/SharingLabel";
-import BlueprintActions from "@/extensionConsole/pages/blueprints/BlueprintActions";
+import InstallableActions from "@/installables/InstallableActions";
 
-import { type InstallableViewItem } from "@/extensionConsole/pages/blueprints/blueprintsTypes";
+import { type InstallableViewItem } from "@/installables/blueprintsTypes";
 import Status from "@/extensionConsole/pages/blueprints/Status";
 import { ListGroup } from "react-bootstrap";
 import LastUpdatedLabel from "@/extensionConsole/pages/blueprints/labels/LastUpdatedLabel";
@@ -55,7 +55,7 @@ const ListItem: React.VoidFunctionComponent<{
         <Status installableViewItem={installableItem} />
       </div>
       <div className="flex-shrink-0">
-        <BlueprintActions installableViewItem={installableItem} />
+        <InstallableActions installableViewItem={installableItem} />
       </div>
     </ListGroup.Item>
   );
