+++ src/extensionConsole/pages/blueprints/listView/ListView.tsx
@@ -21,8 +21,8 @@ import ListItem from "./ListItem";
 import { VariableSizeList as List } from "react-window";
 import ListGroupHeader from "@/extensionConsole/pages/blueprints/listView/ListGroupHeader";
 import { uuidv4 } from "@/types/helpers";
-import { type BlueprintsPageContentProps } from "@/extensionConsole/pages/blueprints/blueprintsTypes";
 import ListItemErrorBoundary from "@/extensionConsole/pages/blueprints/listView/ListItemErrorBoundary";
+import { BlueprintsPageContentProps } from "@/extensionConsole/pages/blueprints/BlueprintsPageContent";
 
 const ROW_HEIGHT_PX = 90;
 const HEADER_ROW_HEIGHT_PX = 43;
