+++ src/extensionConsole/pages/blueprints/modals/shareModals/PublishRecipeModals.tsx
@@ -26,7 +26,7 @@ import Loader from "@/components/Loader";
 import { useOptionalRecipe } from "@/recipes/recipesHooks";
 import EditPublishContent from "./EditPublishContent";
 import CancelPublishContent from "./CancelPublishContent";
-import { isRecipePendingPublish } from "@/extensionConsole/pages/blueprints/utils/installableUtils";
+import { isRecipePendingPublish } from "@/installables/installableUtils";
 import PublishedContent from "./PublishedContent";
 
 const ModalContentSwitch: React.FunctionComponent = () => {
