+++ src/installables/InstallableIcon.module.scss
