+++ src/installables/blueprintsTypes.ts
@@ -15,7 +15,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { type TableInstance } from "react-table";
 import { type RecipeDefinition } from "@/types/recipeTypes";
 import { type ReactNode } from "react";
 import { type Organization } from "@/types/contract";
@@ -72,10 +71,4 @@ export type InstallableViewItem = {
    * True if the source package is no longer available
    */
   unavailable: boolean;
-};
-
-export type BlueprintsPageContentProps = {
-  tableInstance: TableInstance<InstallableViewItem>;
-  width: number;
-  height: number;
 };
